requirement.txt
Django==3.2.6
djangorestframework==3.12.4
mssql-django==1.0
bcrypt==3.2.0
blinker==1.4
PyJWT==2.1.0